update java version and functions

+++ src/main/java/karn/ashish/type3/CalculatorImpl.java
@@ -1,16 +1,16 @@
 package karn.ashish.type3;
 
 public class CalculatorImpl implements Calculator {
-    public int add(int a, int b){
+    public double add(double a, double b){
         return a+b;
     }
-    public int sub(int a, int b){
+    public double sub(double a, double b){
         return a-b;
     }
-    public int mul(int a, int b){
+    public double mul(double a, double b){
         return a*b;
     }
-    public int div(int a, int b){
+    public double div(double a, double b){
         return a/b;
     }
 }
